+++ include/faint/Variables.h
@@ -6,7 +6,7 @@
 #include <unordered_set>
 #include <vector>
 
-#include <rarexsec/data/Types.h>
+#include <faint/data/Types.h>
 
 namespace analysis {
