+++ src/MuonSelector.cc
@@ -1,4 +1,4 @@
-#include "rarexsec/MuonSelector.h"
+#include "faint/MuonSelector.h"
 
 #include <cmath>
